Enemy Behaviour started
IMPORTANT CHANGES ARE ELABORATED ON IN THE #PROGRAMMING TEXT CHANNEL IN DISCORD

Game scenes are now 3d, and all gameobjects are rotated to appear on the top down 2d camera still. this allows me to use NavMesh instead of writing an entire pathfinding algorithm.

If you change the shape of the level, or if we add different levels, let me know and i will set up the pathfinding.
If you have any questions, ask me